To be

+++ eshop-shippingservice/src/main/java/com/samsungsds/eshop/shipping/ShippingRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ShippingRepository extends JpaRepository<ShippingResult, Long> {
+public interface ShippingRepository extends JpaRepository<ShippingResult, Integer> {
 
     ShippingResult findShippingResultByOrderId(Integer orderId);
 }
